requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-filter==21.1
django-import-export==2.8.0
django-modeltranslation==0.18.2
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
Markdown==3.3.7
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
django-cors-headers==3.13.0
gunicorn==20.1.0
